Set proper names for edit and details pages, and fix the logic.

+++ src/app/games/details/details.component.ts
@@ -1,61 +1,45 @@
 import { Component, OnInit } from '@angular/core';
 import { MaterialLibraryModule } from '../../material-library/material-library.module';
 import { CommonModule } from '@angular/common';
-import { FormBuilder, FormGroup, ReactiveFormsModule } from '@angular/forms';
+import { Game } from '../../types/games';
 import { ActivatedRoute, Router } from '@angular/router';
 import { FirestoreService } from '../../services/firestore.service';
 
 @Component({
   selector: 'app-details',
   standalone: true,
-  imports: [MaterialLibraryModule, CommonModule, ReactiveFormsModule ],
+  imports: [MaterialLibraryModule, CommonModule],
   templateUrl: './details.component.html',
   styleUrl: './details.component.scss'
 })
 export class DetailsComponent implements OnInit {
-  gameForm: FormGroup;
-  gameId!: string;
+  game: Game | null = null;
+  gameId: string;
 
   constructor(
-    private fb: FormBuilder,
     private route: ActivatedRoute,
     private router: Router,
     private firestoreService: FirestoreService
   ) {
-    this.gameForm = this.fb.group({
-      title: [''],
-      year: [''],
-      designer: [''],
-      artist: [''],
-      publisher: [''],
-      rating: [''],
-      category: [''],
-      description: [''],
-      image: ['']
-    });
+    this.gameId = this.route.snapshot.paramMap.get('id')!;
   }
 
   ngOnInit(): void {
-    this.gameId = this.route.snapshot.paramMap.get('id')!;
     this.loadGameDetails();
   }
 
   async loadGameDetails() {
-    const game = await this.firestoreService.getGameById(this.gameId);
-    if (game) {
-      this.gameForm.patchValue(game);
-    }
+    this.game = await this.firestoreService.getGameById(this.gameId);
   }
 
-  async onSave() {
-    if (this.gameForm.valid) {
-      await this.firestoreService.updateGame(this.gameId, this.gameForm.value);
+  async deleteGame() {
+    if (this.gameId) {
+      await this.firestoreService.deleteGame(this.gameId);
       this.router.navigate(['/catalog']);
     }
   }
 
-  async onDelete() {
-    await this.firestoreService.deleteGame(this.gameId);
-    this.router.navigate(['/catalog']);
+  editGame() {
+    this.router.navigate(['/edit', this.gameId]);
   }
 }
